Moved Difficulty enum to its own file
